refactored the javascript to take advantage of es6
