+++ .eslintrc.json
@@ -13,7 +13,13 @@
     "rules": {
       "no-shadow": "off",
       "no-param-reassign": "off",
-      "eol-last": "off"
+      "eol-last": "off",
+      "import/extensions": [
+        "error",
+        {
+            "js": "ignorePackages"
+        }
+    ]
     },
     "ignorePatterns": [
       "dist/",
