+++ index.html
@@ -26,12 +26,13 @@
       </ul>
     </nav>
     <main>
+      <p id="time"></p>
       <div id="books-wrapper" class="books-wrapper">
         <h1>Awesome Books</h1>
-        <section class="book-section" id="book-section">
-        </section>
+
+        <section class="book-section" id="book-section"></section>
       </div>
-      
+
       <form id="form" class="form" action="">
         <h2>Add a new book</h2>
         <input type="text" id="author-input" placeholder="Author" required />
@@ -42,20 +43,19 @@
       <section class="contact-page" id="contact-page">
         <h1 class="contact-header">Contact Information</h1>
         <ul>
-          Do have any questions or you just want to say "Hello"?<br>You can reach out to us!<br><br>
+          Do have any questions or you just want to say "Hello"?<br />You can
+          reach out to us!<br /><br />
           <li>Our e-mail mail@mail.com</li>
-                <li>Our phone number: 0043586534422</li>
-                <li>Our address Streetname 22, 84503 City, Country</li>
+          <li>Our phone number: 0043586534422</li>
+          <li>Our address Streetname 22, 84503 City, Country</li>
         </ul>
       </section>
     </main>
 
     <footer>
-      <p>
-        copyright.
-      </p>
+      <p>copyright.</p>
     </footer>
 
-    <script src="./index.js"></script>
+    <script src="./index.js" type="module"></script>
   </body>
 </html>
