+++ modules/time.js
@@ -1,0 +1,9 @@
+import { DateTime } from '../node_modules/luxon/src/luxon.js';
+
+const currentTime = () => {
+  const time = DateTime.now().toLocaleString(DateTime.DATETIME_FULL);
+  const timeSection = document.getElementById('time');
+  timeSection.innerHTML = time;
+};
+
+export default currentTime;
